Remove dependency on the sqlalchemy model.
